NLP/requirements.txt
numpy
scipy
spacy
textstat
pyyaml
